ES-13 Now handling voice banks with corrupted file paths
UTSU will now scan the filesystem for existing wav and frq files, Shift-JIS decode them and use these as the filenames internally.

This is especially handy when a voice bank has been created on a Japanese locale and installed on a non-Japanese locale.

NOTE: this means we can no longer derive the file names for wav and frq names based upon the lyric name. These must be supplied explicitly.